requirements.txt
certifi==2020.6.20
greenlet==1.1.2
mysql-connector-python==8.0.27
protobuf==3.19.1
PyMySQL==1.0.2
SQLAlchemy==1.4.27
